Fix PE64 image_base truncation and update PE tests with accurate reference values
This commit fixes a critical bug and improves test accuracy:

**Bug Fix: PE64 Image Base Truncation**
- Changed pe_file::image_base() return type: uint32_t -> uint64_t
- Changed pe_file::image_base_ member type: uint32_t -> uint64_t
- Removed truncation in PE64 parsing (line 118 of pe_file.cpp)
- PE32+ files now correctly report 64-bit image base addresses
  - Example: TCMADM64.EXE now reports 0x140000000 (not 0x40000000)

**Test Improvements: Accurate Virtual Size Values**
- Updated test_pe32.cpp section size expectations
  - Used actual VirtualSize from PE section headers
  - dump-pe "Size" shows aligned/padded values (SizeOfRawData)
  - Parser correctly returns unaligned VirtualSize
- Updated test_pe64.cpp section size expectations
  - Same correction for 64-bit PE sections

**Test Results**
- All 36 test cases pass
- All 408 assertions pass
- PE32 (TCMDX32.EXE): 4 sections verified
- PE32+ (TCMADM64.EXE): 5 sections verified (includes .pdata)

+++ include/libexe/pe_file.hpp
@@ -49,7 +49,7 @@ namespace libexe {
             [[nodiscard]] pe_file_characteristics characteristics() const;
 
             /// Optional Header accessors
-            [[nodiscard]] uint32_t image_base() const;
+            [[nodiscard]] uint64_t image_base() const;
             [[nodiscard]] uint32_t entry_point_rva() const;
             [[nodiscard]] uint32_t section_alignment() const;
             [[nodiscard]] uint32_t file_alignment() const;
@@ -85,7 +85,7 @@ namespace libexe {
             uint16_t section_count_ = 0;
             uint32_t timestamp_ = 0;
             uint16_t characteristics_ = 0;
-            uint32_t image_base_ = 0;
+            uint64_t image_base_ = 0;
             uint32_t entry_point_rva_ = 0;
             uint32_t section_alignment_ = 0;
             uint32_t file_alignment_ = 0;
